renaming launch files appropriate to their functions
